requirements.txt
asgiref==3.5.0
beautifulsoup4==4.11.1
bootstrap4==0.1.0
confusable-homoglyphs==3.2.0
Django==4.0.4
django-bootstrap4==22.1
django-registration==3.2
django-tinymce==3.4.0
djangorestframework==3.13.1
Pillow==9.1.0
psycopg2-binary==2.9.3
python-decouple==3.6
pytz==2022.1
soupsieve==2.3.2.post1
sqlparse==0.4.2
